Game mechanic is complete
so i made the whole game mechanic
the answers never go higher then 100 so it will be easier to solve. i even added score system. so i made a class and a mono
the class collects data a gives it to the mono
the mono takes the data and turns it to string and give it to text.

it needs a timer which i will do later